JitArm64: Fix broken format strings in Arm64RegCache

+++ Source/Core/Core/PowerPC/JitArm64/JitArm64_RegCache.cpp
@@ -72,7 +72,8 @@ void Arm64RegCache::LockRegister(ARM64Reg host_reg)
 {
   auto reg = std::find(m_host_registers.begin(), m_host_registers.end(), host_reg);
   ASSERT_MSG(DYNA_REC, reg != m_host_registers.end(),
-             "Don't try locking a register that isn't in the cache. Reg %d", host_reg);
+             "Don't try locking a register that isn't in the cache. Reg %d",
+             static_cast<int>(host_reg));
   reg->Lock();
 }
 
@@ -80,7 +81,8 @@ void Arm64RegCache::UnlockRegister(ARM64Reg host_reg)
 {
   auto reg = std::find(m_host_registers.begin(), m_host_registers.end(), host_reg);
   ASSERT_MSG(DYNA_REC, reg != m_host_registers.end(),
-             "Don't try unlocking a register that isn't in the cache. Reg %d", host_reg);
+             "Don't try unlocking a register that isn't in the cache. Reg %d",
+             static_cast<int>(host_reg));
   reg->Unlock();
 }
